refactor(xrpl): rename nft package with `nft` name
