+++ xrpl/queries/nft/nftoken_buy_offers_test.go
@@ -1,4 +1,4 @@
-package path
+package nft
 
 import (
 	"testing"
