+++ xrpl/queries/nft/types/nftoken_offer.go
@@ -5,10 +5,10 @@ import (
 )
 
 type NFTokenOffer struct {
-	Amount            types.CurrencyAmount `json:"amount"`
-	Flags             uint                 `json:"flags"`
-	NFTokenOfferIndex string               `json:"nft_offer_index"`
-	Owner             types.Address        `json:"owner"`
-	Destination       types.Address        `json:"destination,omitempty"`
-	Expiration        int                  `json:"expiration,omitempty"`
+	Amount            any           `json:"amount"`
+	Flags             uint          `json:"flags"`
+	NFTokenOfferIndex string        `json:"nft_offer_index"`
+	Owner             types.Address `json:"owner"`
+	Destination       types.Address `json:"destination,omitempty"`
+	Expiration        int           `json:"expiration,omitempty"`
 }
